Uplift of #10415 (squashed) to beta
