+++ browser/resources/settings/BUILD.gn
@@ -112,14 +112,13 @@ preprocess_if_expr("preprocess") {
     "brave_overrides/privacy_page.js",
     "brave_overrides/reset_profile_dialog.js",
     "brave_overrides/safety_check_passwords_child.js",
-    "brave_overrides/search_engines_section.js",
+    "brave_overrides/search_page.js",
     "brave_overrides/security_page.js",
     "brave_overrides/settings_menu.js",
     "brave_overrides/settings_section.js",
     "brave_overrides/settings_ui.js",
     "brave_overrides/site_details.js",
     "brave_overrides/site_settings_page.js",
-    "brave_overrides/search_page.js",
     "brave_overrides/sync_account_control.js",
     "brave_overrides/sync_controls.js",
     "brave_reset_page/brave_reset_profile_dialog_behavior.js",
@@ -223,7 +222,6 @@ ts_library("build_ts") {
     "brave_overrides/privacy_page.js",
     "brave_overrides/reset_profile_dialog.js",
     "brave_overrides/safety_check_passwords_child.js",
-    "brave_overrides/search_engines_section.js",
     "brave_overrides/search_page.js",
     "brave_overrides/security_page.js",
     "brave_overrides/settings_menu.js",
