+++ browser/resources/settings/brave_overrides/index.js
@@ -30,7 +30,7 @@ import './printing_page.js'
 import './privacy_page.js'
 import './reset_profile_dialog.js'
 import './safety_check_passwords_child.js'
-import './search_engines_section.js'
+import './search_page.js'
 import './security_page.js'
 import './settings_menu.js'
 import './settings_section.js'
@@ -39,4 +39,3 @@ import './site_details.js'
 import './site_settings_page.js'
 import './sync_account_control.js'
 import './sync_controls.js'
-import './search_page.js'
